producer: remove unused frontierAdvanced parameter
The frontierAdvanced parameter was always set to true. Removing the
option and hard coding the true behavior makes it possible to simplify
the checkpointPacer implementation.

Part of: #130367
Release notes: None
